+++ Changelog.md
@@ -1,6 +1,12 @@
 ### development version
 
-[full changelog](https://github.com/Mange/roadie/compare/v1.0.2...master)
+[full changelog](https://github.com/Mange/roadie/compare/v1.0.3...master)
+
+* Nothing yet
+
+### 1.0.3
+
+[full changelog](https://github.com/Mange/roadie/compare/v1.0.2...v1.0.3)
 
 * Bug fixes
   * Don't change `asset_providers` of a `Roadie::Document` when applying Options with no `asset_providers` set.